Added lazy rendering mode
By specifying '-l', the screen is redrawn only when DXYN is executed.
This increases performance and reduces the risk of memory corruption due
to back-buffer updates coinciding with buffer swaps (either in SDL or in
the backing driver). The ROMs can now be executed at well over 1kHz,
making the experience sharper, with less graphical artifacts.