Reworked display system
Now using SDL scaling in stead of rendering to a Texture and copying the
result to the screen buffer.

Back buffer updates no longer occur when displaying sprites. In stead,
we only update the logical screen state. Operations on the back buffer
are performed on screen refresh.

Picked a new color scheme.